Feature/backend/nom 8 expenses (#43)
* feat(NOM-8): add expenses with participants and payments

* feat(NOM-8): add expense validation for trip membership

* feat(NOM-8): improve activities list filtering and clean controllers

* feat(NOM-8): refactor expense and dto

* feat(NOM-8): quick creatrip fix

* feat(NOM-8): handle optional password fields in user update flow

* feat(NOM-8): quick fix

* feat(NOM-8): implement expenses, balances and debt calculation

* feat(NOM-8): quick response addition

* feat(NOM-8): validate participant removal rules and properly map expense-activity in response

* feat(NOM-8): code cleanup and implement endpoint for trip total calculation

* feat(NOM-8): quick fix on delete trip

* feat(NOM-8):  modularize user login register update logic and validation

* feat(NOM-8):  add automatic debt settlement endpoint

* feat(NOM-8): add user debt progress summary in trip debts endpoint

* feat(NOM-8): payment managing

+++ api/src/main/java/nomadia/Service/ExpenseService.java
@@ -117,9 +117,7 @@ public class ExpenseService {
                             "Monto del split inválido");
                 }
                 if (owedMap.putIfAbsent(user.getId(), split.getAmountOwed()) != null) {
-                    throw new ResponseStatusException(
-                            HttpStatus.BAD_REQUEST,
-                            "Split duplicado para el mismo usuario");
+                    throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Split duplicado para el mismo usuario");
                 }
                 involvedUsers.add(user.getId());
                 totalOwed = totalOwed.add(split.getAmountOwed());
@@ -218,6 +216,9 @@ public class ExpenseService {
         Expense expense = expenseRepository.findByIdWithTrip(expenseId)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Gasto no encontrado"));
         tripService.getTripAndValidateMember(expense.getTrip().getId(), userId);
+        if(paymentRepository.existsByTripId(expense.getTrip().getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"No se pueden eliminar gastos de un viaje cuando existen pagos registrados");
+        }
         expenseRepository.delete(expense);
     }
 
@@ -290,6 +291,7 @@ public class ExpenseService {
             }
         }
     }
+
     public List<DebtDTO> calculateDebts(List<UserBalanceDTO> balances) {
         List<BalanceNode> creditors = balances.stream()
                 .filter(b -> b.getBalance().compareTo(BigDecimal.ZERO) > 0)
